Prototyping #1 - Player Movement & Actions
-Imported Assets-

I imported an asset called "POLYGON Fantasy Kingdom Packby Synty Studios" which will be the placeholder for the environment and the level design for the game.

-Player Movement & Action-

I use the unity package "InputSystem" to create the player movement. With the assistance of ChatGPT, I was able to create the rotation that I needed as it ties in with the attack system that will be the core mechanic of my game. I tested that system and it is exactly the type of system that I plan on implementing.