Changing Checkpoint System
Instead of colliding with the zone that serves as a checkpoint, the new checkpoint is an unlit campfire with a flaming skull hovering above it. Players must now attack the skull to destroy it and light the campfire, which will become their new respawn point, while removing the old respawn point.