Player Character: Spin Jump, Healing Effects
The player character will play a spin animation when pressing the jumping input. I've made the spin animation play to create the idea that the player character needs to do a spin to jump.
Healing Particle plays whenever the player interacts with the healing object or when the player triggers a new checkpoint.